refactor: Move /gallery and /projects enpoints
Now located in a separate file, reach at /api/db/...
